// src/config/interfaces/interfaces.ts
export interface ICategory {
  id: number;
  name: string;
  slug: string;
  meta_title: string;
  watermark: string;
  meta_discriptions: string;
}